fix: unbox NumberSetting Double casts and add HUDMod no-arg accessors
Java forbids narrowing a boxed Double directly to int/float/long, so the
204 (int)/(float)/(long) setting.getValue() sites are unboxed via (double)
first. Adds no-arg getX/getY/getWidth/getHeight to HUDMod so HUD modules
compile against the expected API.

+++ src/main/java/net/glacierclient/modules/advanced/atmosphere/CameraClipTweak.java
@@ -21,7 +21,7 @@ public class CameraClipTweak extends GlacierMod {
 
     @Override
     public void onEnable() {
-        currentNearClip = (float) nearClip.getValue();
+        currentNearClip = (float)(double) nearClip.getValue();
         targetNearClip = currentNearClip;
     }
 
@@ -29,7 +29,7 @@ public class CameraClipTweak extends GlacierMod {
 
     @Override
     public void onTick() {
-        targetNearClip = (float) nearClip.getValue();
+        targetNearClip = (float)(double) nearClip.getValue();
         if (smoothTransitions.getValue()) {
             float lerpSpeed = (float) (transitionSpeed.getValue() * 0.1);
             currentNearClip += (targetNearClip - currentNearClip) * lerpSpeed;
